move formatDate to utils
